feat: compile-time toggle for TRACE and DEBUG only
INFO higher levels always compiled, there is still a run-time level which defaults to warn
