+++ cpp/farm_ng/core/logging/CMakeLists.txt
@@ -14,12 +14,13 @@ farm_ng_add_library(farm_ng_core_logging
     eigen.h
     expected.h
     format.h
-    logger_impl.h
+    trace_debug_log.h
+    logger.h
     backtrace.h
   SOURCES
     format.cpp
     expected.cpp
-    logger_impl.cpp
+    logger.cpp
     backtrace.cpp
 )
 
@@ -34,7 +35,9 @@ foreach(test_basename
     expected
     format
     logger
-    logger_define)
+    logger_define_trace_variant
+    logger_define_debug_variant
+    logger_define_info_variant)
   farm_ng_add_test(${test_basename}
                     PARENT_LIBRARY farm_ng_core_logging
                     LINK_LIBRARIES farm_ng_core_logging
